requirements.txt
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
